% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boundaryTransformation}
\alias{boundaryTransformation}
\title{Linear transformation of parameter vector to fit values to boundaries}
\usage{
boundaryTransformation(boundaries, pop)
}
\arguments{
\item{boundaries}{List of boundary transformation parameters}

\item{pop}{Population for boundary transformation}
}
\value{
Vector of population values after boundary transformation
}
\description{
Linear transformation of parameter vector to fit values to boundaries
}
