% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boundaryTransformationInit}
\alias{boundaryTransformationInit}
\title{Boundary transformation rules init}
\usage{
boundaryTransformationInit(parameters)
}
\arguments{
\item{List}{of $max and 4min values}
}
\value{
List of boundary transformation parameters
}
\description{
Boundary transformation rules init
}
