% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmaesInit}
\alias{cmaesInit}
\title{CMA-ES parameters init}
\usage{
cmaesInit(values, stdDevs, inseed, lambda)
}
\arguments{
\item{values}{Initial population}

\item{stdDevs}{Standart deviations}

\item{inseed}{Initial random seed}

\item{lambda}{Population size}
}
\value{
cmaes struct
}
\description{
CMA-ES parameters init
}
