% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmaesUpdateDistribution}
\alias{cmaesUpdateDistribution}
\title{CMA-ES population distribution update for each iteration}
\usage{
cmaesUpdateDistribution(cmaes, rgFunValVec)
}
\arguments{
\item{cmaes}{Current CMA-ES parameters}

\item{rgFuncValueVec}{cost function values}
}
\value{
New cmaes parameters
}
\description{
CMA-ES population distribution update for each iteration
}

