% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmaesInit}
\alias{cmaesInit}
\title{CMA-ES parameters init}
\usage{
cmaesInit(values, stdDevs, inseed)
}
\arguments{
\item{values}{Initial population}

\item{stdDevs}{Standart deviations}
}
\value{
inseed Initial random seed
}
\description{
CMA-ES parameters init
}

